fix: #133 API 공통 Response 수정
1. API 공통 Response 수정
2. 추가적으로 손봐야 하는 부분이 있음

+++ app/shop/src/main/java/com/shoppingmall/repository/PostMapper.java
@@ -18,8 +18,7 @@ public interface PostMapper {
 
     int savePost(Post post);
 
-    Long updatePostById(Post post);
+    int updatePostById(Post post);
 
-    Long deletePostById(long postId);
-
+    int deletePostById(long postId);
 }
